+++ totalsegmentator/map_to_binary.py
@@ -125,6 +125,11 @@ class_map = {
         1: "body_trunc",
         2: "body_extremities",
     },
+    "pleural_pericard_effusion": {
+        # 1: "lung_pleural",
+        2: "pleural_effusion",
+        3: "pericardial_effusion",
+    },
     "test": {
         1: "carpal",
         2: "clavicula",