+++ totalsegmentator/preview.py
@@ -68,6 +68,9 @@ roi_groups = {
     "body": [
         ["body_trunc", "body_extremities"]
     ],
+    "pleural_pericard_effusion": [
+        ["pleural_effusion", "pericardial_effusion"]
+    ],
     "test": [
         ["carpal", "clavicula", "femur", "fibula", "humerus", "metacarpal", "metatarsal", 
         "patella", "hips", "phalanges_hand", "radius", "ribs", "scapula", "skull", "spine",